% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ukgeo.R
\name{ukgeo}
\alias{ukgeo}
\title{Get data from the UK Geoportal}
\usage{
ukgeo(query_string = NULL)
}
\arguments{
\item{query_string}{A query string to pass to the UK Geoportal}
}
\value{
A data frame of the results
}
\description{
Get data from the UK Geoportal
}
\examples{
u1 = paste0("https://services1.arcgis.com/ESMARspQHYMw9BZ9/",
  "arcgis/rest/services/Combined_Authorities_December_2022_EN_BFC",
  "/FeatureServer/0/query?where=1\%3D1&outFields=*&outSR=4326&f=json")
ukgeo(u1)
}
